+++ core-bundle/CHANGELOG.md
@@ -2,5 +2,6 @@
 
 ### 4.3.0-beta1 (2016-XX-XX)
 
+ * Automatically find the template files. 
  * Make the legacy configuration files optional (see #521).
  * Move the analytics templates to the head section (see contao/core#8182).